add mask owner draw with canvas

+++ src/gfx/gfx_line_generator.h
@@ -12,7 +12,7 @@
 namespace gfx {
 
 // dda_line_interpolator
-template <int FractionShift, int YShift = 0>
+template <int32_t FractionShift, int32_t YShift = 0>
 class gfx_dda_line_interpolator
 {
 public:
@@ -23,9 +23,9 @@ public:
     {
     }
 
-    gfx_dda_line_interpolator(int y1, int y2, unsigned int count)
+    gfx_dda_line_interpolator(int32_t y1, int32_t y2, uint32_t count)
         : m_y(y1)
-        , m_inc(((y2 - y1) << FractionShift) / (int)count)
+        , m_inc(((int32_t)((uint32_t)(y2 - y1) << FractionShift)) / (int32_t)count)
         , m_dy(0)
     {
     }
@@ -40,30 +40,30 @@ public:
         m_dy -= m_inc;
     }
 
-    void operator += (unsigned int n)
+    void operator += (uint32_t n)
     {
         m_dy += m_inc * n;
     }
 
-    void operator -= (unsigned int n)
+    void operator -= (uint32_t n)
     {
         m_dy -= m_inc * n;
     }
 
-    int y(void) const { return m_y + (m_dy >> (FractionShift - YShift)); }
-    int dy(void) const { return m_dy; }
+    int32_t y(void) const { return m_y + (m_dy >> (FractionShift - YShift)); }
+    int32_t dy(void) const { return m_dy; }
 
 private:
-    int m_y;
-    int m_inc;
-    int m_dy;
+    int32_t m_y;
+    int32_t m_inc;
+    int32_t m_dy;
 };
 
 // dda2 line interpolator
 class gfx_dda2_line_interpolator
 {
 public:
-    typedef int save_data_type;
+    typedef int32_t save_data_type;
     enum {
         save_size = 2,
     };
@@ -74,7 +74,7 @@ public:
     }
 
     // forward-adjusted line
-    gfx_dda2_line_interpolator(int y1, int y2, int count)
+    gfx_dda2_line_interpolator(int32_t y1, int32_t y2, int32_t count)
         : m_cnt(count <= 0 ? 1 : count)
         , m_lft((y2 - y1) / m_cnt)
         , m_rem((y2 - y1) % m_cnt)
@@ -90,7 +90,7 @@ public:
     }
 
     // backward-adjusted line
-    gfx_dda2_line_interpolator(int y1, int y2, int count, int)
+    gfx_dda2_line_interpolator(int32_t y1, int32_t y2, int32_t count, int)
         : m_cnt(count <= 0 ? 1 : count)
         , m_lft((y2 - y1) / m_cnt)
         , m_rem((y2 - y1) % m_cnt)
@@ -105,7 +105,7 @@ public:
     }
 
     // backward-adjusted line
-    gfx_dda2_line_interpolator(int y, int count)
+    gfx_dda2_line_interpolator(int32_t y, int32_t count)
         : m_cnt(count <= 0 ? 1 : count)
         , m_lft(y / m_cnt)
         , m_rem(y % m_cnt)
@@ -161,16 +161,16 @@ public:
         m_mod += m_cnt;
     }
 
-    int mod(void) const { return m_mod; }
-    int rem(void) const { return m_rem; }
-    int lft(void) const { return m_lft; }
-    int y(void) const { return m_y; }
+    int32_t mod(void) const { return m_mod; }
+    int32_t rem(void) const { return m_rem; }
+    int32_t lft(void) const { return m_lft; }
+    int32_t y(void) const { return m_y; }
 private:
-    int m_cnt;
-    int m_lft;
-    int m_rem;
-    int m_mod;
-    int m_y;
+    int32_t m_cnt;
+    int32_t m_lft;
+    int32_t m_rem;
+    int32_t m_mod;
+    int32_t m_y;
 };
 
 }
